+++ sscRaft/processing/__init__.py
@@ -1,4 +1,5 @@
 from .alignments import *
+from .filters import *
 from .io import *
 from .opt import *
 from .background_correction import *